fixed display bug at cart

+++ public/js/empty-cart.js
@@ -1,27 +1,25 @@
 const emptyCartButton = document.getElementById('emptyCartButton');
-const eraseItemButton = document.querySelectorAll('eraseItem');
+const eraseItemButtons = document.querySelectorAll('.eraseItemButton');
 
-emptyCartButton.addEventListener('click', async () => {
-
-  const cartId = emptyCartButton.value;
-  const response = await fetch(`http://localhost:8080/api/carts/${cartId}`, {
-    method: 'DELETE'
+if (emptyCartButton) {
+  emptyCartButton.addEventListener('click', async () => {
+    const cartId = emptyCartButton.value;
+    const response = await fetch(`http://localhost:8080/api/carts/${cartId}`, {
+      method: 'DELETE'
+    });
+    if (response.ok) {
+      alert('Cart deleted successfully.');
+      location.reload();
+    } else {
+      alert('Error: Unable to delete the cart.');
+    }
   });
-  if (response.ok) {
-    alert('Cart deleted successfully.');
-    location.reload();
-  } else {
-    alert('Error: Unable to delete the cart.');
-  }
-});
+}
 
-
-document.querySelectorAll('.eraseItemButton').forEach(button => {
-
+eraseItemButtons.forEach(button => {
   button.addEventListener('click', async () => {
-
-    let itemId = button.dataset.itemId;
-    let cartId = button.dataset.cartId;
+    let itemId = button.getAttribute('data-item-id');
+    let cartId = button.getAttribute('data-cart-id');
 
     const response = await fetch(`http://localhost:8080/api/carts/${cartId}/products/${itemId}`, {
       method: 'DELETE'
@@ -35,7 +33,3 @@ document.querySelectorAll('.eraseItemButton').forEach(button => {
     }
   });
 });
-
-
-
-
